clean up old code and prep
